refactor : Mentor 로 통일 및 폴더구조 개선했습니다
