+++ src/app/mentor/_components/MentorClient/index.tsx
@@ -1,6 +1,6 @@
 import MenteePageTabs from "./_components/MenteePageTabs";
 import MentorFindSection from "./_components/MentorFindSection";
-import MentoPageTabs from "./_components/MentorPageTabs";
+import MentorPageTabs from "./_components/MentorPageTabs";
 import MyMentorSection from "./_components/MyMentorSection";
 
 const MentorClient = () => {
@@ -11,7 +11,7 @@ const MentorClient = () => {
         // 멘토페이지
         <>
           {/* 나의 멘토 - 멘티 탭 및 채팅카드 */}
-          <MentoPageTabs />
+          <MentorPageTabs />
           {/* 나의 멘토 페이지 */}
           <MyMentorSection />
         </>
@@ -20,8 +20,6 @@ const MentorClient = () => {
         <>
           {/* 탭 및 나의 멘토 , 멘티요청 리스트 채팅카드 */}
           <MenteePageTabs />
-          {/* 중간 밑줄 */}
-          <div className="mb-10 mt-10 h-1.5 w-full bg-k-50"></div>
           {/* 멘토찾기 섹션 */}
           <MentorFindSection />
         </>
